% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiVplot.R
\name{multiVplot}
\alias{multiVplot}
\title{Multivariate Plot}
\usage{
multiVplot(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
plot
}
\description{
This is the beginning of a plot function
}
\details{
This function will be editd and added to through a community of R programmers to make plots suitable for Multivariate statistics
}
\examples{
df = data.frame(x= 1:30, y = (1:30)^2)
 mulitVplot(df)
}
